+++ Tests/RichTextKitTests/Data/RichTextDataFormatTests.swift
@@ -12,6 +12,18 @@ import XCTest
 
 class RichTextDataFormatTests: XCTestCase {
 
+    let vendorType = UTType(exportedAs: "fooType")
+
+    lazy var vendorFormat = RichTextDataFormat.vendorArchivedData(
+        id: "foo",
+        fileExtension: "fooFile",
+        uniformType: vendorType
+    )
+
+    func idResult(for format: RichTextDataFormat) -> String {
+        format.id
+    }
+
     func formatResult(for format: RichTextDataFormat) -> [RichTextDataFormat] {
         format.convertableFormats
     }
@@ -29,27 +41,43 @@ class RichTextDataFormatTests: XCTestCase {
     }
 
 
+    func testLibraryFormatsReturnAllNonVendorFormats() {
+        XCTAssertEqual(RichTextDataFormat.libraryFormats, [.archivedData, .plainText, .rtf])
+    }
+
+
+    func testConvertableIdIsValidForAllFormats() {
+        XCTAssertEqual(idResult(for: .archivedData), "archivedData")
+        XCTAssertEqual(idResult(for: .plainText), "plainText")
+        XCTAssertEqual(idResult(for: .rtf), "rtf")
+        XCTAssertEqual(idResult(for: vendorFormat), "foo")
+    }
+
     func testConvertableFormatsAreValidForAllFormats() {
-        XCTAssertEqual(formatResult(for: .archivedData), [.rtf, .plainText])
+        XCTAssertEqual(formatResult(for: .archivedData), [.plainText, .rtf])
         XCTAssertEqual(formatResult(for: .plainText), [.archivedData, .rtf])
         XCTAssertEqual(formatResult(for: .rtf), [.archivedData, .plainText])
+        XCTAssertEqual(formatResult(for: vendorFormat), [.plainText, .rtf])
     }
 
     func testStandardFileExtensionIsValidForAllFormats() {
         XCTAssertEqual(extensionResult(for: .archivedData), "rtk")
         XCTAssertEqual(extensionResult(for: .plainText), "txt")
         XCTAssertEqual(extensionResult(for: .rtf), "rtf")
+        XCTAssertEqual(extensionResult(for: vendorFormat), "fooFile")
     }
 
     func testSupportsImagesIsValidForAllFormats() {
         XCTAssertTrue(imageResult(for: .archivedData))
         XCTAssertFalse(imageResult(for: .plainText))
         XCTAssertFalse(imageResult(for: .rtf))
+        XCTAssertEqual(extensionResult(for: vendorFormat), "fooFile")
     }
 
     func testUniformTypeIsValidForAllFormats() {
         XCTAssertEqual(uniformTypeResult(for: .archivedData), .archivedData)
         XCTAssertEqual(uniformTypeResult(for: .plainText), .plainText)
         XCTAssertEqual(uniformTypeResult(for: .rtf), .rtf)
+        XCTAssertEqual(uniformTypeResult(for: vendorFormat), vendorType)
     }
 }