+++ Tests/RichTextKitTests/Data/RichTextDataReaderTests.swift
@@ -12,42 +12,21 @@ import XCTest
 
 class RichTextDataReaderWriterTests: XCTestCase {
 
-    func testCanGenerateAndParseArchivedData() {
-        let string = NSAttributedString(string: "foo bar baz")
-        guard let data = try? string.richTextArchivedData() else { return XCTFail("Test failed") }
-        let result = try? NSAttributedString(archivedData: data)
-        XCTAssertEqual(result?.string, "foo bar baz")
-    }
-
-    func testCanGenerateAndParsePlainTextData() {
-        let string = NSAttributedString(string: "foo bar baz")
-        guard let data = try? string.richTextPlainTextData() else { return XCTFail("Test failed") }
-        let result = try? NSAttributedString(plainTextData: data)
-        XCTAssertEqual(result?.string, "foo bar baz")
-    }
-
-    func testCanGenerateAndParseRtfData() {
-        let string = NSAttributedString(string: "foo bar baz")
-        guard let data = try? string.richTextRtfData() else { return XCTFail("Test failed") }
-        let result = try? NSAttributedString(rtfData: data)
-        XCTAssertEqual(result?.string, "foo bar baz")
-    }
-
-    func testDataForFormatCanGenerateAndParseArchivedData() {
+    func testCanGenerateDataForArchivedFormat() {
         let string = NSAttributedString(string: "foo bar baz")
         guard let data = try? string.richTextData(for: .archivedData) else { return XCTFail("Test failed") }
         let result = try? NSAttributedString(data: data, format: .archivedData)
         XCTAssertEqual(result?.string, "foo bar baz")
     }
 
-    func testDataForFormatCanGenerateAndParsePlainTextData() {
+    func testCanGenerateDataForPlainTextFormat() {
         let string = NSAttributedString(string: "foo bar baz")
         guard let data = try? string.richTextData(for: .plainText) else { return XCTFail("Test failed") }
         let result = try? NSAttributedString(data: data, format: .plainText)
         XCTAssertEqual(result?.string, "foo bar baz")
     }
 
-    func testDataForFormatCanGenerateAndParseRtfData() {
+    func testCanGenerateDataForRtfFormat() {
         let string = NSAttributedString(string: "foo bar baz")
         guard let data = try? string.richTextData(for: .rtf) else { return XCTFail("Test failed") }
         let result = try? NSAttributedString(data: data, format: .rtf)