+++ Tests/RichTextKitTests/RichTextCoordinator+SubscriptionsTests.swift
@@ -7,10 +7,9 @@
 //
 
 #if os(iOS) || os(macOS) || os(tvOS)
+import RichTextKit
 import SwiftUI
 import XCTest
-
-@testable import RichTextKit
 
 class RichTextCoordinator_SubscriptionsTests: XCTestCase {
     
@@ -111,9 +110,9 @@ class RichTextCoordinator_SubscriptionsTests: XCTestCase {
     }
 
 
-    func testSelectedRangeUpdatesTextView() {
+    func testSelectedRangeChangeUpdatesTextView() {
         let range = NSRange(location: 4, length: 3)
-        textContext.selectedRange = range
+        textContext.selectRange(range)
         XCTAssertEqual(textView.selectedRange, range)
     }